Implement Stage 2: Add status bar and reset button
Add classes to add a status bar for messages at bottom as well as a
button to reset the field.
Change Cell.java to add common properties and methods to set/get the
current text of the buttons.
Re-factor TicTacToe.java to discrete methods for setting the window
properties and adding components in.