Implement Stage 2: Full implementation
The hardest part for this stage was implementing the action listeners
and being able to access the buttons and status bar in order to change
them as necessary. This solution does work, but is probably overly
contrived, and it may be condensed (reducing the number of classes - is
it REALLY necessary to have each element in its own class??) in further
stages. Definitely wasn't a fan of essentially having the Controller
class be completely static, but it worked, and I didn't see any other
solution at the time. This, too, will probably get re-worked in later
stages.

TicTacToe.java:
- We now assemble all the components in this class and add it to the
  JFrame. Once everything is assembled and added, we pass the two
  components we need to update to the Controller class.

Board.java:
- Moved cells and addCells() to TicTacToe.java.
- Moved clearCells(), checkState(), isThreeInARow(), and INDICES to
  Controller.java.

Cell.java:
- New constructor parameter for the action listener.
- Set the cell text to an initial state in the constructor.

StatusBar.java:
- Update constructor to use Status enum for initial message.

ResetButton.java:
- New constructor parameter for the action listener.

Status.java:
- New enum class for the status messages to display.

Controller.java:
- New constructor to take arguments for the buttons and status bar
  objects we need to manipulate.
- playCell() now correctly only allows a play if the cell is empty.
  Also updates the status while in play and determines when the game is
  drawn.
- checkState() updates the status when a winner is found and updates the
  gameOver boolean.
- isThreeInARow() now ensures that we only return true if we've matched
  X or O.
- Move the actionPerformed() into an inner class.