Implement Stage 4: Update status messages during game
Update the status messages during the game to now show who is playing.
Also, the winning message is more descriptive.

Small problem with handling menu clicks - due to how we've implemented
everything, starting a game results in an incorrect state. This will
need to be fixed, probably by restructuring the program and/or
changing the Controller's constructor. We'll see in the next commit.

TicTacToe.java:
- was naming wrong menu item - now setting name on the menu instead of
  the menu bar

Controller.java:
- add MessageFormat's to be used in status bar
- field move now starts at 0 and increments -- eases getting button text
  from the correct player button
- update playCell() to handle change to field move as well as logic flow
  change
- update checkState() to use new win message
- handleMenuItem() now correctly handles the menu clicks